requirements.txt
numpy
pytorch-lightning==2.0.9